fixed better better better

+++ src/OrderBeverageService/OrderBeverageHandler.h
@@ -83,7 +83,7 @@ namespace vending_machine{
       //3. get beverage name
       std::string beverageName ="";
       try {
-        beverageName = getbeverage_client->ReturnBeverage(weatherType);
+        beverageName = getbeverage_client->ReturnBeverage((int)weatherType);
       } 
       catch (...) {
         _getbeverage_client_pool->Push(getbeverage_client_wrapper);
